Fix: libpacemaker: Anti-colocations affect scores when roles match
Currently, we skip an anti-colocation if the role specified in the
constraint for the dependent or primary resource doesn't match the
actual next role of the dependent or primary resource. This is the
opposite of the correct behavior. The anti-colocation should apply only
if both the dependent's next role and the primary's next role match the
specified next role for each (if set) in the colocation.

This hasn't caused a problem in and of itself up to now. This is
likely because an earlier block handles promoted/unpromoted role
constraints when the dependent is already assigned, as is the case when
choosing roles for the dependent.

Ref T672

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>