Test: scheduler: Fix incorrect promotion scores in promoted-probed-score
The promotion score attribute names refer to the specific clone
instances rather than to the primitive (AdminDrbd):

<nvpair
 id="status-hypatia-corosync.nevis.columbia.edu-promoted-AdminDrbd.0"
 name="master-AdminDrbd:0"
 value="5"
/>

<nvpair
 id="status-orestes-corosync.nevis.columbia.edu-promoted-AdminDrbd.1"
 name="master-AdminDrbd:1"
 value="5"
/>

As a result, if each instance gets assigned to the other node (in this
case, if instance 0 gets assigned to orestes and instance 1 to hypatia),
each instance's assigned node has no promotion score for it. The
instances cannot promote.

Currently, the scores change slightly, and the transition does not.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>