Refactor: libpacemaker: Drop pcmk__add_collective_constraints()
This logic was added by commit 2be9cc39 as a way to avoid unnecessary
shuffling of instances of clones that are involved in colocations. It
has grown a bit over time.

However, our recent instance shuffling fix (for
clone-recover-no-shuffle-{4..7} tests) made all of this obsolete. It's
now simpler to drop pcmk__add_collective_constraints() and
can_run_everywhere(), and to add parent constraints to instances
indiscriminately. The code that applies colocations and adds dependent
colocated node scores can handle them, and the instance assignment code
is more robust.

Note that we drop the "has influence" check that was in
pcmk__add_collective_constraints(), but that gets checked later.

Ref T489

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>