Low: python: No longer use a default section argument.
Previously when modifying the constraints section of the generated CIB,
we would not pass any section argument.  This would mean we'd default to
"all", which is not a valid section as far as cibadmin is concerned.

Instead, pass "constraints" for the section here.  This gets rid of the
only place where we were using the default, so it can be made into a
required parameter.