Test: scheduler: Update after fixing anticolocation roles
The previous commit fixed an issue in which anticolocations affected
scores when the role did NOT match, rather than when the role DID match.

Regarding the test changes: currently, bundle colocations apply only to
the bundle itself and to its containers. Bundle containers are not
promotable. So, a colocation with role=Promoted will never match for a
bundle container. For that reason, before the fix, colocations with
role=Promoted affected the bundle containers.

Now that a colocation affects scores only when the resource role matches
the colocation role, the constraints in these tests no longer have an
effect.

Ref T672

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>