API: libcrmcommon: xml_top argument of xpath_search() is now const
xml_top->doc is passed as a non-const variable to xmlXPathNewContext(),
but xmlXPathNewContext() doesn't modify it. It's just another libxml2
function that's unnecessarily non-const.

For clarity of intent, we typically declare a Pacemaker variable as non-
const even if the variable itself is not modified, as long as one of its
members is modified via that variable. However, convenience wins here
and it doesn't break any C rules.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>