Low: python: Default to the latest CIB schema version.
Changing how a ConfigFactory instance was created introduced a bug in
validating the CIB schema.

Before that patch, if you asked for a "pacemaker30" CIB, that would get
translated to "pacemaker-3.7" by the CIB30.version attribute.  This
version attribute would in turn be passed on the "cibadmin --empty"
command line, which would end up being the string written into the CIB's
validate-with property.  That "pacemaker-3.7" string directly maps to
the xml/pacemaker-3.7.rng file.

After the patch, we're not doing all that translation any more.
Instead, we're just using whatever the value from the Schema key in the
Environment.  This is set by default to "pacemaker-3.0" when adding all
the command line arguments.  This is too old of a schema version for a
lot of what we're doing now.

So instead, just grab whatever the most recent schema version is from
the build options and use that as the default all around.  The user can
still specify an older version on the command line, and if you really
need additional support for something older, there's always checking out
a different git branch.