Refactor: python: Unindent a bunch of code blocks.
I didn't do them all, only the ones that seem to benefit most from this
change.  Basically, invert the test condition and continue/return.  Then
un-indent everything that was previously indented under the original
condition.

This transformation could probably be performed on a lot of other code.
I've not been very consistent on making this change.