Fix: tools: Fix a bug in clone resource description display
Previously, descriptions of resources running on multiple
nodes were displayed despite --full not being used (with pcs
status) or --show-detail not being used (with crm_mon).

For example, clone resources running on multiple nodes were
affected.

Now, --full and --show-detail must be used in order for resource
descriptions to be displayed, regardless of the number of nodes
the resource is run on.

see bz: 2106642