Low: daemons: Disable sync points in certain circumstances.
This only affects sync points on IPC messages containing multiple
update messages, and only when those messages are received by an attrd
that does not support multiple update messages.  Hopefully, this is a
pretty small number of installations.

In these cases, right now we act like sync points are supported but they
are not.  For local sync points, what happens is we iterate over each
individual update message.  Each update message will have a sync point
attribute, and the server will send the ACK for each as it processes the
message.  However, the client is only going to wait for the first ACK
before continuing.  For all the other ACKs, there will not be a client
listening.

Cluster-wide sync points likely work the same way, though I have not
tested it to be sure.  I don't see how they could work, at least.

I don't see any way to actually handle these, so my "fix" here is to
simply disable support.  Log a message and remove the attributes from
the XML messages so the server doesn't try to do anything.

Fixes T601