Low: scheduler: make validate_on_fail() replace value
Previously, validate_on_fail() validated the on-fail meta-attribute, and
returned it if valid, or an appropriate fallback if not.

Now, if on-fail is not valid, directly replace it in the meta-attributes table.

While this could theoretically have side effects, since previously the table
retained the raw value (and added it to the transition graph), nothing actually
needs the raw value.