Test: cts-scheduler: update expected outputs for on-fail change
on-fail="stop" is not valid for stop actions, and it is now (properly) removed
from transition graphs (though the controller doesn't care, so it has no real
effect).

Cleanups now (properly) have on-fail="ignore" in the transition graph, but
again the controller doesn't care, so it has no real effect.