Fix: build: Fix API and CIB versions in RPM build
When running configure from within the RPM build directory, the relative
paths that set API_VERSION and CIB_VERSION don't match anything in the
git index. However, the git ls-files command returns success even if its
output is empty. So we never fall back to the ls command if the
git ls-files output is empty. As a result, the API_VERSION and
CIB_VERSION variables get set to empty strings in an RPM build.

This causes cts-lab to fail in an RPM installation. The validate-with
attribute in the empty CIB is set based on the value of CIB_VERSION. If
CIB_VERSION is empty, validate-with gets set to "pacemaker-", which is
invalid.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>