Low: controller: Ignore CIB replace notification during shutdown
attrd now uses a CIB transaction to update attribute values. A
successful commit-transaction request replaces the CIB. So when attrd
sets the shutdown attribute for a node, the CIB gets replaced.

If the node is the DC, this triggers a new election via the
do_cib_replaced() callback. We shouldn't trigger an election when we're
shutting down.

This issue was observed via the following log message while running
cts-lab:

error: Input I_STOP received in state S_FINALIZE_JOIN from notify_crmd

It only happened once in 100 runs, so this will be sanity-only.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>