Low: controller: Drop CIB replace call recording and forgetting
All we really care about is whether we (the local controller instance)
requested a particular operation that replaced all or part of the CIB.
If we requested it, then we shouldn't restart the join sequence simply
because we're notified that based tried to perform the operation. If we
do want to restart the join sequence after a replace that we requested,
we can do so via a request-specific callback.

I'm not sure why I created a table to record each call in the first
place... probably because the previous code set R_CIB_ASKED only for
specific calls. We don't need to be selective anymore, now that each
CIB client has a unique client ID that's included in the notify XML.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>