Test: python: Specify a path to python tests.
In a source checkout, the Makefiles and all the tests are in the same
directory.

However, we do support building into a separate directory from the
source and in fact this is what CI does.  On these systems, Makefiles
are built from Makefile.am and therefore they end up in the build
directory.  For that matter, any tests whose names end with .in would
also end up in the build directory.  The rest of the tests stay in the
source directory.

We need to check both of these places for unit tests to run.

We also need to add both to PYTHONPATH so python can import built python
files as well as ones that don't need to be built.

Note that for the moment, we do not have any built unit tests.  We might
as well be ready for the future, however.

Fixes T693