Refactor: based, libcib: Drop all CIB replace notify code
Nothing internal uses it anymore, and we've never advertised it to
external callers. cib_types.h doesn't document valid values for the
event argument of cib_api_operations_t:add_notify_callback(). Any
external caller using this was relying on an undocumented feature or
implementation detail and can't expect it to work.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>