Build: configure: process datarootdir correctly
autoconf wants .in files to ensure datarootdir is set, because directories such
as datadir are defined in terms of it. However, we expand all paths before
substitution, so it doesn't apply to us. Setting AC_DATAROOTDIR_CHECKED tells
autoconf not to worry about it, but for completeness, we should expand
datarootdir as well.