Refactor: libcrmcommon: Use libgcrypt for hashing.
Aside from some annoying library setup, libgcrypt supplies a hashing
function that is just about as easy to use as md5_buffer that we were
getting from gnulib.  Plus, it's a shared library that we are already
shipping (and linking against, somehow) with its own upstream that we
don't have to bundle and update in our source.

My lame unit test passes, as do all the regression and scheduler tests.