Build: Lists in Makefiles should be more consistent.
The idea here is that each item of a list is on its own line.  Each item
is tabbed over until it's under the previous item, with spaces to round
out the padding.  The backslash is tabbed over until it's right of
whatever the longest element in the list is.  For some lists that only
have a couple items of very similar length, I've just let the editor
line them up with spaces.

Set your tabstop to 8 whenever editing a Makefile to preserve this
indentation.

Fixes T250