Refactor: tools: Remove direct calls to crm_perror.
There's very few of these calls anymore.  However, the tools still make
plenty of calls to library functions that in turn call crm_perror.
Those are going to be much more difficult to replace, but luckily
they're being tracked in T172 so we don't need to deal with them here.

Fixes T171