Doc: Pacemaker Explained: formalize CIB properties table
This starts a new model for options tables: a new table for configuration value
types defines allowed values, and the options tables will have columns for
option name (indexed, and labeled so it can be linked from elsewhere),
option type (linked to types table), option default, and option description.

The CIB properties is the first (and currently only) table converted to the new
model. Additionally, use rST's list-table instead of table to be able to sanely
edit it.