Feature: scheduler: raise node-pending-timeout default to 2h
There is a valid use case for an administrator to troubleshoot cluster issues
by manually starting Corosync and investigating before starting Pacemaker.
In that case, the node will remain pending.

However, while a node is pending, the cluster is unable to manage resources,
due to being unable to probe resources on the pending node. This makes choosing
a default a balancing act. Two hours seems like a decent trade-off, and of
course the administrator can change it at any time.