Feature: daemons: Download newer schema files to a remote executor.
If the remote executor supports an older version of the CIB schema than
the rest of the cluster, various operations could fail due to the schema
not validating on the remote node.

Instead, ask the CIB manager for the updated schema files, store them in
a separate directory on the remote node, and add them to the list of
known schema files.  These can then be used just like the packaged
schema files.

Note that while this is a fairly large patch, it's really just a lot of
boring IO and file management code, which requires a lot of error
handling.

Fixes T182