Refactor: libcrmcommon: Add pcmk__load_schemas_from_dir.
This just moves the bulk of crm_schema_init out into its own function,
allowing us to call it multiple times as other schema directories
appear.

There is no similar function for unloading schemas from a directory.  If
you want to do that, you'll have to use crm_schema_cleanup to unload
everything and start over.

This function has not been tested for the possibility that the same
schema files exist in multiple directories.  It is assumed that it won't
be used like that.