Refactor: scheduler: replace pe_order_optional internally
... with pcmk__ar_ordered.

The name is questionable. The relation is optional if only this flag is set
(see should_add_input_to_graph()), and effectively this enables the relation
by making it not match an equality comparison against pcmk__ar_none. But
enabling it means order_actions() will order them, and this name is readable
and in the future could more specifically control whether the relation is an
ordering.