Refactor: scheduler: add enum pcmk__action_relation_flags
... initially with no values.

This new enum in libcrmcommon is intended to replace enum pe_ordering from
libpe_status and libpe_rules, which will be deprecated as public API. No
external code should need that level of detail into the ordering code, and
nearly everything that uses the enum is already internal anyway.

All internal references to enum pe_ordering are updated to the new one. The
reference in pe_action_wrapper_t is left since it is public API (it should
really be uint32_t anyway).

The "ordering" terminology is replaced here with "action relation" since
non-ordering relationships are also involved.