Test: cts-lab: work around issues with Reattach test
Previously, the Reattach test sometimes failed due to resource activity
that was expected.

Work around that by disabling stonith-class and oc2b-type resources in the test
setup, and re-enabling them in teardown. Also reprobe all resources as part of
setup, to ensure node-attribute-based rules are in effect before the test.

This allows us to drop previous workarounds of not running the test when oc2b
resources were present, and checking for cluster stability multiple times in a
row.

Fixes T709