Refactor: scheduler: simplify searching for action config
All of the callers pcmk__find_action_config() know the action name and
interval, so pass those instead of the action key. This is much simpler and
more efficient.

This also allows more accurate detection of migration actions (previously, it
could wrongly use the "migrate" configuration if a resource had "migrate_to" or
"migrate_from" in its name).