Doc: pacemakerd: Mark PCMK_shutdown_delay as deprecated
It's undocumented and the string constant is internal-only, so this only
requires COMPAT comments.

The variable was added by f4ae91a to debug shutdown issues when
Pacemaker is managed by systemd. That's a narrow use case and it
probably hasn't been useful for a long time.

See comment at
https://github.com/ClusterLabs/pacemaker/pull/3245#discussion_r1372029854.

Ref T637

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>