Fix: scheduler: don't show pending nodes as having "<3.15.1" feature set
When a node's controller completes the join process, it sets the #feature-set
node attribute, which is used in status displays to detect mixed-version
clusters.

Previously, the status displays used "<3.15.1" if the node was online (joined
at the cluster layer) but had no #feature-set attribute. However, this will
also be the case when the node is pending (coming up between when it has joined
the cluster layer and completed the join process, and going down between when
it has left the controller group and the cluster layer).

Now, show "<3.15.1" only if the node has completed the join process.