Refactor: libcrmcommon: Rename xml_minimum_schema_index.
This function's name is unclear.  It actually returns the most recent
X.0 schema index.  The new name is pretty bad, but I think it's at least
clear.

And then while I'm at it, rewrite it to make it more clear.  Just
iterate the known_schemas list, looking for the right .0 one.  This code
does not get run very often, and it caches its result, so there's no
need to do the reverse traversal with a lagging index.