Refactor: libcrmcommon: Change how schema upgrade versions are handled
...in update_validation.  Schemas always either upgrade to the next one
in the list, or do not upgrade.  The latter only happens when we get to
the last real version and the next one is pacemaker-next/none.

With that change made, we also need to change the conditional.  There's
no need to check that the upgrade will regress.  We only need to check
that we've run off the end of the list of real schema versions.

A future commit will remove the after_transform variable entirely, but
since this is its most visible and complicated use, splitting this into
a separate commit seems worth it.