Refactor: libcrmcommon: Get rid of after_transform.
As stated in the previous commit, schemas always just upgrade to the
next one in the list.  There's no need to keep track of that fact, so
get rid of the variable that held it.  This then allows us to get rid of
all the places that value was being set and passed around.