Feature: daemons: Only ask for schemas if supported by the server
We can use LRMD_PROTOCOL_VERSION in the handshake to determine what the
server supports, similar to what is being done in attrd.  Add a macro to
compare the version we were given with the known minimum version that
supports the schema transfer command.

Additionally, introduce LRMD_COMPATIBLE_PROTOCOL which is just the major
version number required for the connection to succeed.  This gets rid of
the need for LRMD_MIN_PROTOCOL_VERSION, which can now be deprecated.

And then since I wasn't sure compare_version would work if you give it a
full version number and just a major version, add a unit test for that.