Test: libpacemaker: Add unit tests for pcmk_resource_delete.
This also introduces the first unit tests for any of the public
functions in libpacemaker.  We can use this as a template for other
public functions.  Also note that at least in this case, the private
function is well tested by the combination of the public function unit
test plus cts-cli.