Refactor: libcrmcommon: Remove unnecessary schema code.
The block that sets the version if we didn't previously do so is no
longer necessary.  This block only executes if the version parameter is
all zeros, which at the moment is only "pacemaker-next" and "none".  We
could probably guarantee this will continue to be the case.

Additionally, I don't see that this would even do anything useful
anymore.  Scanning the name for a version number is going to fail for
"pacemaker-next" and "none".  So really, this block was just handling
the possibility that we passed in no version number but that the name
contained a number.

And with that done, there's only one more spot using schema_scanf so we
can just replace that with a call to sscanf.