Refactor: libcrmcommon: pcmk__xml_output_finish should take an exit code.
It previously did, but that was removed because the exit code was also
being used to determine whether the function should do any output or
not.

However, this means that the public libpacemaker API functions will
write out an XML block with a <status code=""> wrapper that always looks
successful.  This is because pcmk__xml_output_finish previously called
out->finish with 0 for the exit code always.  The actual exit code of
the process was not taken into account.