Doc: books: Move clone RA requirements to Pacemaker Administration
As a general rule, we document information relevant to writing agents in
Pacemaker Administration, and we document information relevant to
behavior and configuration in Pacemaker Explained. The boundary is quite
fuzzy but Admin seems like a better fit for these and will give us a
convenient place to add some upcoming documentation.

Also convert tables to list-tables for easier editing.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>