Test: libcrmcommon: Rename pcmk__xml_find_x_0_schema_index_test.
The function was renamed to pcmk__find_x_0_schema_index, so the unit
test needs to be renamed to match.  This doesn't affect the test being
run at all, but test files and test functions need to be named the same
in order for my coverage checking script to count the function as
tested.